Add initial implementation of a virtual list
Despite the browser optimizations there is a limit in the number of
elements that can be added to a document before the browser becomes
sluggish when the document is further modified.

This problem can be clearly seen in the list of chat messages with long
conversations; adding several messages to the document all at once
instead of one by one can improve the performance of the initial load,
but when the number of messages grows too much any other addition causes
the browser to become unresponsive. In this case adding several messages
at once (which may not be possible either during a regular chat) does
not help, because the real culprit here is that the browser needs to
layout again all the previous messages.

To solve that problem this commit introduces an initial implementation
of a virtual list; the virtual list keeps in the document only those
elements that are currently visible and refreshes them as needed when
the list is scrolled.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>